refactor: 재료 목록 조회 API 수정
- id값을 long타입으로 변경
- status 컬럼 조회 제거
- 재료 조회 시 유저가 등록하거나 기본으로 제공되는 것 중 비공개되지 않은 항목만 조회하도록 로직 변경
- dto 전달명 변경

+++ src/main/java/com/recipe/app/src/fridgeBasket/application/dto/FridgeBasketDto.java
@@ -14,7 +14,7 @@ public class FridgeBasketDto {
     @AllArgsConstructor
     @NoArgsConstructor(access = AccessLevel.PRIVATE)
     public static class FridgeBasketIdsRequest {
-        private List<Integer> ingredientList;
+        private List<Long> ingredientList;
     }
 
     @Getter
